+++ app/App.tsx
@@ -1,0 +1,7 @@
+import { Button } from 'components/ui/Button'
+
+function App() {
+  return <Button variant='login'>Registration</Button>
+}
+
+export default App
