+++ app/components/ui/Button.tsx
@@ -1,0 +1,39 @@
+import { cva, type VariantProps } from 'class-variance-authority'
+import { cn } from 'lib/utils'
+import { forwardRef, type ButtonHTMLAttributes } from 'react'
+
+const buttonVariants = cva('', {
+  variants: {
+    variant: {
+      default: '',
+      registration: 'w-[334px] h-[49px] rounded-lg bg-black text-white',
+      login: 'w-[130px] h-[49px] rounded-lg bg-transparent text-black',
+      outline: 'border border-input bg-background shadow-sm hover:bg-accent hover:text-accent-foreground',
+      secondary: 'bg-secondary text-secondary-foreground shadow-sm hover:bg-secondary/80',
+      ghost: 'hover:bg-accent hover:text-accent-foreground',
+      link: 'text-primary underline-offset-4 hover:underline',
+    },
+    size: {
+      default: 'h-9 px-4 py-2',
+      sm: 'h-8 rounded-md px-3 text-xs',
+      lg: 'h-10 rounded-md px-8',
+      icon: 'h-9 w-9',
+    },
+  },
+  defaultVariants: {
+    // variant: 'default',
+    // size: 'default',
+  },
+})
+
+type ButtonProps = ButtonHTMLAttributes<HTMLButtonElement> & VariantProps<typeof buttonVariants>
+
+export const Button = forwardRef<HTMLButtonElement, ButtonProps>(
+  ({ className, variant, size, children, ...props }, ref) => (
+    <button className={cn(buttonVariants({ variant, size, className }))} ref={ref} {...props}>
+      {children}
+    </button>
+  )
+)
+
+Button.displayName = 'Button'
