+++ index.html
@@ -8,6 +8,6 @@
   </head>
   <body class="font-poppins text-fs-14-lh-normal-fw-500">
     <div id="root"></div>
-    <script type="module" src="src/index.tsx"></script>
+    <script type="module" src="app/index.tsx"></script>
   </body>
 </html>
