--- src/App.tsx
@@ -1,5 +1,0 @@
-function App() {
-  return <p className='bg-sidebar text-white/50'>Hello, world!</p>
-}
-
-export default App
